test: split slow zero page tests (#17514)
